requirements.txt
pytest
requests
requests-ratelimiter
networkx
psycopg2-binary
setuptools
wheel
build 